feat: added run request function, repstructured the place
