+++ src/hooks/collections/collection.ts
@@ -70,3 +70,5 @@ export const useUpdateCollection = (workspaceId: string) => {
     },
   });
 };
+
+
